DEVEX-293 rethrow NoClassDefFound errors within #main
In case of bug a NoClassDefFoundError occurred with ExportOperation of FirstSpirit (called by export command)
cause the used fs-isolated-runtime.jar was incompatible to the used FirstSpirit server.
This NoClassDefFoundError was never seen because of special handling within Main#main to indicate missing api jar.
From now on all other occurrences of NoClassDefFoundError will be rethrown.